+++ R/utils.R
@@ -77,3 +77,34 @@ explode.tbl_kgx <- function(graph, ...) {
   })
   return(graphs)
 }
+
+
+
+#' Recursively Merge Two Named Lists
+#'
+#' This function takes two named lists and merges them recursively. Values in
+#' the second list (`b`) override or extend those in the first list (`a`).
+#' If both `a` and `b` contain a named list at the same position, they are merged recursively.
+#'
+#' @param a A named list to be updated.
+#' @param b A named list containing values to override or add to `a`.
+#'
+#' @return A named list that is the result of merging `b` into `a`.
+#' @export
+#'
+#' @examples
+#' a <- list(x = list(y = 1, z = 2), foo = "bar")
+#' b <- list(x = list(y = 42, new = 99), foo = "baz", extra = "new_value")
+#' merge_lists(a, b)
+merge_lists <- function(a, b) {
+	for (name in names(b)) {
+		if (is.list(b[[name]]) && name %in% names(a) && is.list(a[[name]])) {
+			# Recursively merge if both elements are lists
+			a[[name]] <- merge_lists(a[[name]], b[[name]])
+		} else {
+			# Otherwise, override/add the value from b to a
+			a[[name]] <- b[[name]]
+		}
+	}
+	return(a)
+}